chore: move `KeyboardNavigation` to `common` first
